refactor: rename jax2stablehlo to jax_cc and update references across the codebase
